Add is_following and is_subscribed fields to feed, reorg logic
