+++ src/app/pcasts/controllers/__init__.py
@@ -9,7 +9,8 @@ from app.pcasts.dao import users_dao, \
   recommendations_dao, \
   followings_dao, \
   listening_histories_dao, \
-  itunes_dao
+  itunes_dao, \
+  feed_dao
 
 from app.pcasts.utils.authorize import *
 from app.pcasts.models._all import *
