+++ src/app/pcasts/controllers/get_feed_controller.py
@@ -1,16 +1,5 @@
-import time
-import heapq
+from app.pcasts.dao.feed_dao import FeedContexts
 from . import *
-
-class Enum(set):
-  def __getattr__(self, name):
-    if name in self:
-      return name
-    raise AttributeError
-
-FeedContexts = Enum(['FOLLOWING_RECOMMENDATION',
-                     'FOLLOWING_SUBSCRIPTION',
-                     'NEW_SUBSCRIBED_EPISODE'])
 
 class GetFeedController(AppDevController):
 
@@ -26,96 +15,24 @@ class GetFeedController(AppDevController):
     maxtime = request.args['maxtime']
     page_size = int(request.args['page_size'])
 
-    following_recommendations = \
-      followings_dao.get_following_recommendations(user.id, maxtime, page_size)
-    following_subscriptions = \
-      followings_dao.get_following_subscriptions(user.id, maxtime, page_size)
-    new_subscribed_episodes = \
-      subscriptions_dao.get_new_subscribed_episodes(user.id, maxtime, page_size)
+    feed = feed_dao.get_feed(user, maxtime, page_size)
+    serialized_feed = [
+        feed_dao.attach_fields_to_json(f, self.serialize(f), user) for f in feed
+    ]
 
-    # TODO - huge hack, fix later
-    augmented_episodes = []
-    for e in new_subscribed_episodes:
-      e.created_at = e.pub_date
-      augmented_episodes.append(e)
+    return {'feed': serialized_feed}
 
-    feed = self.merge_sorted_feed_sources(
-        [
-            following_recommendations,
-            following_subscriptions,
-            augmented_episodes
-        ],
-        [
-            FeedContexts.FOLLOWING_RECOMMENDATION,
-            FeedContexts.FOLLOWING_SUBSCRIPTION,
-            FeedContexts.NEW_SUBSCRIBED_EPISODE
-        ],
-        page_size)
-
-    return {'feed': [f.serialize() for f in feed]}
-
-  @classmethod
-  def merge_sorted_feed_sources(cls, sources, contexts, page_size):
-    """
-    Given a list of sources from which to generate a feed page, this function
-      merges the lists together in descending order of when the elements were
-      created.
-
-    Args:
-        sources (list): A list where each element is a list of models containing
-          information for a feed element. Each sublist must already be sorted
-          by descending time.
-        contexts (list): A list where contexts[i] is the `FeedContexts` type
-          of sources[i].
-        page_size (int): Desired length of feed page.
-
-    Returns:
-        list: A list of `FeedElements` of length
-          min(page_size, sum([len(source) for source in sources])) that is
-          sorted by descending time created.
-    """
-    sources = [list(reversed(s)) for s in sources]  # reverse so we can .pop()
-    result, heap, count = [], [], 0
-    for source_idx, s in enumerate(sources):
-      if s:
-        item = s.pop()
-        heapq.heappush(heap, (-time.mktime(item.created_at.timetuple()),
-                              item, source_idx))
-    while len(heap) > 0 and count < page_size:  #pylint: disable=C1801
-      _, item, source_idx = heapq.heappop(heap)
-      result.append(FeedElement(item, contexts[source_idx]))
-      if sources[source_idx]:
-        raw_content = sources[source_idx].pop()
-        heapq.heappush(heap, (-time.mktime(raw_content.created_at.timetuple()),
-                              raw_content, source_idx))
-      count += 1
-    return result
-
-class FeedElement(object):
-
-  def __init__(self, raw_content, context):
-    self.context = context
-    self.time = int(time.mktime(raw_content.created_at.timetuple()))
-    if context == FeedContexts.FOLLOWING_RECOMMENDATION:
-      self.context_supplier = raw_content.user
-      self.content = raw_content.episode
-    elif context == FeedContexts.FOLLOWING_SUBSCRIPTION:
-      self.context_supplier = raw_content.user
-      self.content = raw_content.series
-    elif context == FeedContexts.NEW_SUBSCRIBED_EPISODE:
-      self.context_supplier = raw_content.series
-      self.content = raw_content
-
-  def serialize(self):
+  def serialize(self, feed_element):
     context_to_schemas = {
         FeedContexts.FOLLOWING_RECOMMENDATION: (user_schema, episode_schema),
         FeedContexts.FOLLOWING_SUBSCRIPTION: (user_schema, series_schema),
         FeedContexts.NEW_SUBSCRIBED_EPISODE: (series_schema, episode_schema),
     }
-    supplier_schema, content_schema = context_to_schemas[self.context]
+    supplier_schema, content_schema = context_to_schemas[feed_element.context]
     return {
-        "context": self.context,
-        "time": self.time,
-        "context_supplier": supplier_schema.dump(self.context_supplier).data,
-        "content": content_schema.dump(self.content).data
+        "context": feed_element.context,
+        "time": feed_element.time,
+        "context_supplier": supplier_schema.\
+            dump(feed_element.context_supplier).data,
+        "content": content_schema.dump(feed_element.content).data
     }
